feat(api): hasura trigger & actions refactoring
