+++ .env.build
@@ -21,6 +21,7 @@ GRAPHQL_SERVER_URI=http://localhost:5000/v1/graphql
 API_URL_DOCKER=http://localhost:4000
 WHATEVER_URL=http://localhost:4001/graphql
 HASURA_WEB_HOOK_BASE_URL=http://localhost:4000
-WEB_HOOK_URL_EMAIL_RESERVATION=http://localhost:4000/webhook/email-reservation
-WEB_HOOK_URL_EMAIL_CANCEL_RESERVATION=http://localhost:4000/webhook/email-cancel-reservation
-WEB_HOOK_URL_EMAIL_ACCOUNT_VALIDATION=http://localhost:4000/webhook/email-account-validation
+HASURA_WEB_HOOK_SECRET=z1VJ23P54jQF9JEU5lXJaSVwpuM2vr9NT3ZAuesCNFFx7hrs5F8jrdv16Z3apCaN
+WEB_HOOK_URL_EMAIL_RESERVATION=http://localhost:4000/hasura/events/email-reservation
+WEB_HOOK_URL_EMAIL_CANCEL_RESERVATION=http://localhost:4000/hasura/events/email-cancel-reservation
+WEB_HOOK_URL_EMAIL_ACCOUNT_VALIDATION=http://localhost:4000/hasura/events/email-account-validation
