+++ .env.sample
@@ -15,11 +15,12 @@ GRAPHQL_SERVER_URI=http://localhost:5000/v1/graphql
 API_URL_DOCKER=http://host.docker.internal:4000
 WHATEVER_URL=http://host.docker.internal:4001/graphql
 HASURA_WEB_HOOK_BASE_URL=http://host.docker.internal:4000
-WEB_HOOK_URL_EMAIL_RESERVATION=http://host.docker.internal:4000/webhook/email-reservation
-WEB_HOOK_URL_EMAIL_CANCEL_RESERVATION=http://host.docker.internal:4000/webhook/email-cancel-reservation
-WEB_HOOK_URL_EMAIL_ACCOUNT_VALIDATION=http://host.docker.internal:4000/webhook/email-account-validation
-WEB_HOOK_URL_EMAIL_SERVICE_MEMBER_INVITATION=http://host.docker.internal:4000/webhook/email-service-member-invitation
-WEB_HOOK_URL_TOKEN_REQUEST=http://host.docker.internal:4000/webhook/token-request
+HASURA_WEB_HOOK_SECRET=z1VJ23P54jQF9JEU5lXJaSVwpuM2vr9NT3ZAuesCNFFx7hrs5F8jrdv16Z3apCaN
+WEB_HOOK_URL_EMAIL_RESERVATION=http://host.docker.internal:4000/hasura/events/email-reservation
+WEB_HOOK_URL_EMAIL_CANCEL_RESERVATION=http://host.docker.internal:4000/hasura/events/email-cancel-reservation
+WEB_HOOK_URL_EMAIL_ACCOUNT_VALIDATION=http://host.docker.internal:4000/hasura/events/email-account-validation
+WEB_HOOK_URL_EMAIL_SERVICE_MEMBER_INVITATION=http://host.docker.internal:4000/hasura/events/email-service-member-invitation
+WEB_HOOK_URL_TOKEN_REQUEST=http://host.docker.internal:4000/hasura/events/token-request
 JWT_KEY=abc120d67
 DATABASE_URL=psql://emjpm:test@db/emjpm
 SMTP_HOST=in-v3.mailjet.com
