+++ docker-compose.yaml
@@ -56,6 +56,7 @@ services:
       HASURA_GRAPHQL_UNAUTHORIZED_ROLE: 'anonymous'
       WHATEVER_URL: $WHATEVER_URL
       HASURA_WEB_HOOK_BASE_URL: $HASURA_WEB_HOOK_BASE_URL
+      HASURA_WEB_HOOK_SECRET: $HASURA_WEB_HOOK_SECRET
       WEB_HOOK_URL_EMAIL_RESERVATION: $WEB_HOOK_URL_EMAIL_RESERVATION
       WEB_HOOK_URL_EMAIL_CANCEL_RESERVATION: $WEB_HOOK_URL_EMAIL_CANCEL_RESERVATION
       WEB_HOOK_URL_EMAIL_ACCOUNT_VALIDATION: $WEB_HOOK_URL_EMAIL_ACCOUNT_VALIDATION
