+++ packages/api/package.json
@@ -35,6 +35,7 @@
     "objection": "^2.1.3",
     "objection-db-errors": "^1.1.2",
     "passport": "^0.4.1",
+    "passport-custom": "^1.1.1",
     "passport-http-bearer": "^1.0.1",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
