+++ packages/api/src/routes/hasura-actions/mesures-import/actionsMesuresImporter.js
@@ -1,16 +1,16 @@
-const excelParser = require("../../../../utils/file/excelParser");
-const logger = require("../../../../utils/logger");
+const excelParser = require("../../../utils/file/excelParser");
+const logger = require("../../../utils/logger");
 
-const { Service } = require("../../../../models/Service");
+const { Service } = require("../../../models/Service");
 const actionsMesuresImporterGeoRepository = require("./repository/actionsMesuresImporterGeoRepository");
 const actionsMesuresImporterMesureRepository = require("./repository/actionsMesuresImporterMesureRepository");
-const { Mandataire } = require("../../../../models/Mandataire");
+const { Mandataire } = require("../../../models/Mandataire");
 const actionsMesuresImporterSchemaValidator = require("./schema/actionsMesuresImporterSchemaValidator");
-const { Tis } = require("../../../../models/Tis");
-const { Mesure } = require("../../../../models/Mesure");
+const { Tis } = require("../../../models/Tis");
+const { Mesure } = require("../../../models/Mesure");
 
-const configuration = require("../../../../env");
-const inseeAPI = require("../../../insee-api");
+const configuration = require("../../../env");
+const inseeAPI = require("../../../utils/insee-api");
 
 const actionsMesuresImporter = {
   importMesuresFile
