+++ packages/graphql-server/package.json
@@ -26,7 +26,7 @@
   "devDependencies": {
     "@graphql-codegen/cli": "^1.13.5",
     "@graphql-codegen/introspection": "1.13.5",
-    "@graphql-codegen/typescript": "1.13.5",
+    "@graphql-codegen/typescript": "1.13.5",  
     "@graphql-codegen/typescript-resolvers": "1.13.5",
     "@socialgouv/tslint-config-recommended": "^0.21.0",
     "@types/graphql": "^14.2.3",
